Refactored DB mocks directories
